% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\docType{data}
\name{mockCharacteristics}
\alias{mockCharacteristics}
\title{Mock characteristics results.}
\format{
`Sumary characteristics` summarised result tibble
}
\usage{
mockCharacteristics
}
\description{
Mock characteristics results.
}
\examples{
  mockCharacteristics
}
\keyword{datasets}
