% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\docType{data}
\name{mockCohortDetails}
\alias{mockCohortDetails}
\title{Mock cohort details results.}
\format{
`Cohort details` summarised result tibble
}
\usage{
mockCohortDetails
}
\description{
Mock cohort details results.
}
\examples{
  mockCohortDetails
}
\keyword{datasets}
