% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\docType{data}
\name{mockIncidence}
\alias{mockIncidence}
\title{Mock incidence results.}
\format{
`Incidence estimates` summarised result tibble
}
\usage{
mockIncidence
}
\description{
Mock incidence results.
}
\examples{
  mockIncidence
}
\keyword{datasets}
