% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\docType{data}
\name{mockPrevalence}
\alias{mockPrevalence}
\title{Mock prevalence results.}
\format{
`Prevalence estimates` summarised result tibble
}
\usage{
mockPrevalence
}
\description{
Mock prevalence results.
}
\examples{
  mockPrevalence
}
\keyword{datasets}
