% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockResults.R
\docType{data}
\name{mockSnapshot}
\alias{mockSnapshot}
\title{Mock snapshot results.}
\format{
`CDM snapshot` summarised result tibble
}
\usage{
mockSnapshot
}
\description{
Mock snapshot results.
}
\examples{
  mockSnapshot
}
\keyword{datasets}
