% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsCollection.R
\name{newResultsCollection}
\alias{newResultsCollection}
\title{Create a results_collection object}
\usage{
newResultsCollection(resultsList, types)
}
\arguments{
\item{resultsList}{List of results.}

\item{types}{Specification of the result class.}
}
\value{
A results_collection object
}
\description{
Create a results_collection object
}
