% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsCollection.R
\name{print.results_collection}
\alias{print.results_collection}
\title{Print a results_collection object}
\usage{
\method{print}{results_collection}(x, ...)
}
\arguments{
\item{x}{A results_collection object}

\item{...}{Included for compatibility with generic. Not used.}
}
\value{
Invisibly returns the input
}
\description{
Print a results_collection object
}
