% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Read OMOP CDM files from a path}
\usage{
readFiles(resultsPath)
}
\arguments{
\item{resultsPath}{Path that points to a folder with the result files.}
}
\value{
All elements that could be identified from the csv files.
}
\description{
Read OMOP CDM files from a path
}
\examples{
\dontrun{
library(ShinyGenerator)
readFiles(here::here("Results"))
}
}
