% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFromResults.R
\name{shinyFromResults}
\alias{shinyFromResults}
\title{Create a ShinyApp with the specified results.}
\usage{
shinyFromResults(resultsPath, shinyPath = here::here(), zipFile = "shiny")
}
\arguments{
\item{resultsPath}{Path that points to a folder with the result files.}

\item{shinyPath}{Path where the shiny will be saved.}

\item{zipFile}{Name of the shiny app and the zip file.}
}
\description{
Create a ShinyApp with the specified results.
}
