% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyGenerator-package.R
\docType{package}
\name{ShinyGenerator-package}
\alias{ShinyGenerator}
\alias{ShinyGenerator-package}
\title{ShinyGenerator: Generate Automatic Shiny Apps from OMOP CDM results}
\description{
This package alows the user to generate a shiny app in zip format from OMOP cdm results.
}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
